fix(app-page-builder): expose a useDeleteElement hook
